VMO-7017 do not export interfaces from .vue to avoid >> "export '...' was not found in './Core_SetContactPropertyStore'
